Distance map has been rewritten into integers.

+++ Auticka/src/cz/mff/cuni/autickax/constants/Misc.java
@@ -11,7 +11,7 @@ public final class Misc {
 	/** Determines the color of a pathway */
 	public final Color PATHWAY_COLOR = new Color(0.75f, 0.7f, 0.6f, 1f);
 	/** How many pictures should be blended from the border of pathway */
-	public final float PATHWAY_BORDER_BLEND_DISTANCE = 3;
+	public final int PATHWAY_BORDER_BLEND_DISTANCE = 3;
 	/**
 	 * How big should be the circle in front of start and behind finish in the
 	 * picture of pathway
